style: Refactor all the animated properties to use the style system properly
